The walltime is technically a long. Fixed a possible error with very large walltimes

+++ src/org/griphyn/vdl/karajan/monitor/processors/coasters/BlockRequestedProcessor.java
@@ -61,7 +61,7 @@ public class BlockRequestedProcessor extends AbstractRemoteLogProcessor {
             p.markTo(",");
             int coresPerWorker = Integer.parseInt(p.getToken());
             p.skip("walltime=");
-            int walltime = Integer.parseInt(p.remaining());
+            long walltime = Long.parseLong(p.remaining());
             
             CoasterStatusItem item = (CoasterStatusItem) state.getItemByID(CoasterStatusItem.ID, StatefulItemClass.MISC);
             item.newBlock(blockId, cores, coresPerWorker, walltime);
